Split module
